pyproject.toml
[tool.coverage.run]
command_line = "-m pytest"
omit = [ ".venv/*", "tests/*", "**/*pb2*"]
source = [ ".",]

[tool.black]
extend_exclude = "saturn_sync/plugin|saturn_sync/protocol"
line-length = 88

[tool.isort]
profile = "black"
multi_line_output = 3

# No types are available
# https://github.com/grpc/grpc/issues/29041
[[tool.mypy.overrides]]
module = "grpc"
ignore_missing_imports = true

[[tool.mypy.overrides]]
module = "grpc_health.*"
ignore_missing_imports = true

[tool.poetry]
name = "saturn-sync"
version = "0.0.0"
description = "Synchronize and refactor repositories with ease."
authors = ["Markus Meyer <hydrantanderwand@gmail.com>"]
license = "MPL-2.0"
readme = "README.md"
repository = "https://github.com/wndhydrnt/saturn-sync-protocol"
classifiers = [
    "Development Status :: 3 - Alpha",
    "License :: OSI Approved :: Mozilla Public License 2.0 (MPL 2.0)"
]
packages = [
  { include = "saturn_sync" }
]

[tool.poetry.dependencies]
python = "^3.9"
grpcio = "^1.62.0"
grpcio-health-checking = "^1.62.0"

[tool.poetry.group.dev.dependencies]
black = "^24.2.0"
isort = "^5.13.2"
mypy = "^1.8.0"
types-protobuf = "^4.24.0.20240129"
flake8 = "^7.0.0"
pytest = "^8.1.1"

[tool.poetry.group.dev.dependencies.coverage]
extras = [ "toml",]
version = "7.4.3"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
